Gestión de usuarios y autenticación
Creación de roles, configuración de seguridad de spring security, validaciones de inicio de sesión y creación de usuarios

+++ src/main/java/com/Leucosis/app/service/IUserService.java
@@ -6,7 +6,7 @@ import com.Leucosis.app.entity.User;
 
 public interface IUserService {
 
-	public User getUser(int idUser);
+	public User getUser(Long idUser);
 	public void addUser(User user);
 	public List<User> getAllUser();
 }
